# src/apidocs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "API Aplikasi Penjadwalan Matakuliah",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/profile": {
      "get": {
        "summary": "Profile",
        "tags": [
          "Auth"
        ],
        "operationId": "Profile",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/NoToken1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "no token provided!"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Vary": {
                "type": "string",
                "default": "Origin"
              },
              "Access-Control-Allow-Header": {
                "type": "string",
                "default": "authorization, Origin, Content-Type, Accept"
              },
              "Content-Length": {
                "type": "string",
                "default": "32"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"20-7ZxkElwq5xcc5pY9DGW1oDebWi8\""
              },
              "Set-Cookie": {
                "type": "string",
                "default": "connect.sid=s%3A8tjC8aIXGA54Eaki1T-6z4pktZxSqxGp.wtV6fKFu2UCGQzuTKJVEuqSiVXDV8siE6G6C6YYdo7c; Path=/; HttpOnly"
              },
              "Date": {
                "type": "string",
                "default": "Wed, 15 Jun 2022 13:43:48 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Sukses1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Profil Berhasil Ditampilkan",
                "data": {
                  "id": 1,
                  "nim": "021203",
                  "nama": "Fauzein Mulya Warman",
                  "jurusan": {
                    "id": 1,
                    "nama_jurusan": "S1 Rekayasa Perangkat Lunak"
                  },
                  "ruang": {
                    "id": 1,
                    "nama_kelas": "RPL A",
                    "id_jurusan": 1
                  }
                },
                "status": "mahasiswa"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Vary": {
                "type": "string",
                "default": "Origin"
              },
              "Access-Control-Allow-Header": {
                "type": "string",
                "default": "authorization, Origin, Content-Type, Accept"
              },
              "Content-Length": {
                "type": "string",
                "default": "240"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"f0-aRurEWVcLAivPgVZ1HSzs3nq0Dw\""
              },
              "Date": {
                "type": "string",
                "default": "Wed, 15 Jun 2022 13:45:09 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/matkul-mahasiswa/mahasiswa": {
      "get": {
        "summary": "Matkul Mahasiswa",
        "tags": [
          "API Mahasiswa"
        ],
        "operationId": "MatkulMahasiswa",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "definitions": {
...
    "NoToken1": {
      "title": "NoToken1",
      "example": {
        "message": "no token provided!"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": [
        "message"
      ]
    },
    "Sukses1": {
      "title": "Sukses1",
      "example": {
        "message": "Profil Berhasil Ditampilkan",
        "data": {
          "id": 1,
          "nim": "021203",
          "nama": "Fauzein Mulya Warman",
          "jurusan": {
            "id": 1,
            "nama_jurusan": "S1 Rekayasa Perangkat Lunak"
          },
          "ruang": {
            "id": 1,
            "nama_kelas": "RPL A",
            "id_jurusan": 1
          }
        },
        "status": "mahasiswa"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/Data"
        },
        "status": {
          "type": "string"
        }
      },
      "required": [
        "message",
        "data",
        "status"
      ]
    },
    "Data": {
      "title": "Data",
      "example": {
        "id": 1,
        "nim": "021203",
        "nama": "Fauzein Mulya Warman",
        "jurusan": {
          "id": 1,
          "nama_jurusan": "S1 Rekayasa Perangkat Lunak"
        },
        "ruang": {
          "id": 1,
          "nama_kelas": "RPL A",
          "id_jurusan": 1
        }
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "nim": {
          "type": "string"
        },
        "nama": {
          "type": "string"
        },
        "jurusan": {
          "$ref": "#/definitions/Jurusan"
        },
        "ruang": {
          "$ref": "#/definitions/Ruang"
        }
      },
      "required": [
        "id",
        "nim",
        "nama",
        "jurusan",
        "ruang"
      ]
    },
    "Jurusan": {
      "title": "Jurusan",
      "example": {
        "id": 1,
        "nama_jurusan": "S1 Rekayasa Perangkat Lunak"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "nama_jurusan": {
          "type": "string"
        }
      },
      "required": [
        "id",
        "nama_jurusan"
      ]
    },
    "Ruang": {
      "title": "Ruang",
      "example": {
        "id": 1,
        "nama_kelas": "RPL A",
        "id_jurusan": 1
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "nama_kelas": {
          "type": "string"
        },
        "id_jurusan": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "id",
        "nama_kelas",
        "id_jurusan"
      ]
    },
...
  },
...
}